chore: update customer & customerInfo
