--- src/main/java/com/dangol/dangolsonnimbackend/customer/repository/CustomerInfoRepository.java
@@ -1,8 +1,0 @@
-package com.dangol.dangolsonnimbackend.customer.repository;
-
-import com.dangol.dangolsonnimbackend.customer.domain.CustomerInfo;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CustomerInfoRepository extends JpaRepository<CustomerInfo, Long> {
-
-}
