+++ src/main/java/com/dangol/dangolsonnimbackend/customer/repository/RefreshTokenRepository.java
@@ -1,0 +1,14 @@
+package com.dangol.dangolsonnimbackend.customer.repository;
+
+import com.dangol.dangolsonnimbackend.customer.domain.RefreshToken;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
+
+    Optional<RefreshToken> findByRefreshToken(String refreshToken);
+
+}
