+++ src/main/java/com/dangol/dangolsonnimbackend/customer/service/impl/CustomerServiceImpl.java
@@ -1,11 +1,9 @@
 package com.dangol.dangolsonnimbackend.customer.service.impl;
 
 import com.dangol.dangolsonnimbackend.customer.domain.Customer;
-import com.dangol.dangolsonnimbackend.customer.domain.CustomerInfo;
-import com.dangol.dangolsonnimbackend.customer.dto.CustomerSignupRequestDTO;
-import com.dangol.dangolsonnimbackend.customer.repository.CustomerInfoRepository;
+import com.dangol.dangolsonnimbackend.customer.dto.CustomerSignupDTO;
 import com.dangol.dangolsonnimbackend.customer.repository.CustomerRepository;
-import com.dangol.dangolsonnimbackend.customer.repository.dsl.CustomerInfoQueryRepository;
+import com.dangol.dangolsonnimbackend.customer.repository.dsl.CustomerQueryRepository;
 import com.dangol.dangolsonnimbackend.customer.service.CustomerService;
 import com.dangol.dangolsonnimbackend.errors.BadRequestException;
 import com.dangol.dangolsonnimbackend.errors.enumeration.ErrorCodeMessage;
@@ -17,29 +15,26 @@ import javax.transaction.Transactional;
 public class CustomerServiceImpl implements CustomerService {
 
     private final CustomerRepository customerRepository;
-    private final CustomerInfoRepository customerInfoRepository;
-    private final CustomerInfoQueryRepository customerInfoQueryRepository;
+    private final CustomerQueryRepository customerQueryRepository;
 
-    public CustomerServiceImpl(CustomerRepository customerRepository, CustomerInfoRepository customerInfoRepository, CustomerInfoQueryRepository customerInfoQueryRepository) {
+    public CustomerServiceImpl(CustomerRepository customerRepository, CustomerQueryRepository customerQueryRepository) {
         this.customerRepository = customerRepository;
-        this.customerInfoRepository = customerInfoRepository;
-        this.customerInfoQueryRepository = customerInfoQueryRepository;
+        this.customerQueryRepository = customerQueryRepository;
     }
 
     @Transactional
-    public void signup(CustomerSignupRequestDTO dto) {
+    public void signup(CustomerSignupDTO dto) {
 
         validateDuplicateCustomerInfo(dto);
 
         customerRepository.save(new Customer(dto));
-        customerInfoRepository.save(new CustomerInfo(dto));
     }
 
-    private void validateDuplicateCustomerInfo(CustomerSignupRequestDTO dto) {
-        if (customerInfoQueryRepository.existsByNickname(dto.getNickname())) {
+    private void validateDuplicateCustomerInfo(CustomerSignupDTO dto) {
+        if (customerQueryRepository.existsByNickname(dto.getNickname())) {
             throw new BadRequestException(ErrorCodeMessage.ALREADY_EXISTS_NICKNAME);
         }
-        if (customerInfoQueryRepository.existsByPhoneNumber(dto.getPhoneNumber())) {
+        if (customerQueryRepository.existsByPhoneNumber(dto.getPhoneNumber())) {
             throw new BadRequestException(ErrorCodeMessage.ALREADY_EXISTS_PHONE_NUMBER);
         }
     }
